requirements.txt
Django==3.1.5
djangorestframework
markdown
django-filter
djoser
django-cors-headers
httpie
validate-email